requirements.txt
# Backend Requirements
fastapi==0.109.0
uvicorn==0.24.0
pydantic==2.6.1
google-generativeai==0.3.1
SpeechRecognition==3.10.0
python-multipart==0.0.9

# Frontend Requirements
streamlit==1.30.0
requests==2.31.0
sounddevice==0.4.6
soundfile==0.12.1

# Translation and Audio
gtts==2.5.0